fix: Issue with large nonce not checked
On Hanndshake peers checked to larger nonce
than previously known nonce to them, but they were not
checking and controlling larger nonce, hence accepting
every edge as long as the nonce was increasing. While
this is not fundamentally wrong it can be abused,
provoking overflow and eventually crashing the node.

Test plan
=========
This behavior was created in
- tests/spec/network/controlled_edge_nonce

Now this test should pass